Add QuantileLoss
