+++ CHANGELOG.md
@@ -28,6 +28,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
     - `AUC`
 - `Multinomial` to `proba`
 - `HedgeRegressor` and `StackingBinaryClassifier` to `ensemble`
+- `QuantileLoss` to `optim`
 
 ### Modified
