+++ creme/optim/__init__.py
@@ -8,16 +8,17 @@ from .ada_grad import AdaGrad
 from .adam import Adam
 from .base import Optimizer
 from .ftrl import FTRLProximal
-from .losses import AbsoluteLoss
-from .losses import BinaryClassificationLoss
-from .losses import CauchyLoss
-from .losses import CrossEntropy
-from .losses import EpsilonInsensitiveHingeLoss
-from .losses import HingeLoss
-from .losses import LogLoss
-from .losses import MultiClassificationLoss
-from .losses import RegressionLoss
-from .losses import SquaredLoss
+from .losses.absolute import AbsoluteLoss
+from .losses.base import BinaryClassificationLoss
+from .losses.base import MultiClassificationLoss
+from .losses.base import RegressionLoss
+from .losses.cauchy import CauchyLoss
+from .losses.cross_entropy import CrossEntropy
+from .losses.hinge import EpsilonInsensitiveHingeLoss
+from .losses.hinge import HingeLoss
+from .losses.log_loss import LogLoss
+from .losses.quantile import QuantileLoss
+from .losses.squared import SquaredLoss
 from .lr_schedule import ConstantLR
 from .lr_schedule import InverseScalingLR
 from .lr_schedule import OptimalLR
@@ -47,6 +48,7 @@ __all__ = [
     'NesterovMomentum',
     'OptimalLR',
     'Optimizer',
+    'QuantileLoss',
     'RegressionLoss',
     'RMSProp',
     'SquaredLoss',
