+++ creme/optim/losses/__init__.py
@@ -1,0 +1,11 @@
+__all__ = [
+    'AbsoluteLoss',
+    'BinaryClassificationLoss',
+    'CauchyLoss',
+    'CrossEntropy',
+    'HingeLoss',
+    'LogLoss',
+    'MultiClassificationLoss',
+    'RegressionLoss',
+    'SquaredLoss'
+]
