+++ docs/user-guide.rst
@@ -9,3 +9,4 @@ User guide
    notebooks/building-a-simple-time-series-model.ipynb
    notebooks/the-art-of-using-pipelines.ipynb
    notebooks/debugging-a-pipeline.ipynb
+   notebooks/quantile-regression-uncertainty.ipynb
